requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai
chromadb
pypdf
python-dotenv
tiktoken
crewai